pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Elenlith <mishchenkov@list.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
django-bootstrap4 = "^23.1"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
whitenoise = "^6.4.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
